+++ module/search/search.php
@@ -92,9 +92,11 @@ class search extends common {
 
 		// Mise à jour 2.2
 		if (version_compare($versionData, '2.2', '<') ) {
-			// Déplacer les données du dossier Pages
-			$this->copyDir(self::DATADIRECTORY . 'pages/' . $this->getUrl(0), self::DATADIRECTORY . $this->getUrl(0));
-			$this->removeDir(self::DATADIRECTORY . 'pages/');
+			if (is_dir(self::DATADIRECTORY . 'pages/')) {
+				// Déplacer les données du dossier Pages
+				$this->copyDir(self::DATADIRECTORY . 'pages/' . $this->getUrl(0), self::DATADIRECTORY . $this->getUrl(0));
+				$this->removeDir(self::DATADIRECTORY . 'pages/');
+			}
 			// Mettre à jour la version
 			$this->setData(['module',$this->getUrl(0),'config', 'versionData', '2.2' ]);
 		}
